+++ modules/ingester/local_block.go
@@ -1,4 +1,4 @@
-package wal
+package ingester
 
 import (
 	"context"
@@ -16,9 +16,9 @@ import (
 
 const nameFlushed = "flushed"
 
-// LocalBlock is a block stored in a local storage.  It can be searched and flushed to a remote backend, and
+// localBlock is a block stored in a local storage.  It can be searched and flushed to a remote backend, and
 // permanently tracks the flushed time with a special file in the block
-type LocalBlock struct {
+type localBlock struct {
 	common.BackendBlock
 	reader backend.Reader
 	writer backend.Writer
@@ -26,11 +26,12 @@ type LocalBlock struct {
 	flushedTime atomic.Int64 // protecting flushedTime b/c it's accessed from the store on flush and from the ingester instance checking flush time
 }
 
-var _ common.Finder = (*LocalBlock)(nil)
+var _ common.Finder = (*localBlock)(nil)
 
-func NewLocalBlock(ctx context.Context, existingBlock common.BackendBlock, l *local.Backend) (*LocalBlock, error) {
+// newLocalBlock creates a local block
+func newLocalBlock(ctx context.Context, existingBlock common.BackendBlock, l *local.Backend) *localBlock {
 
-	c := &LocalBlock{
+	c := &localBlock{
 		BackendBlock: existingBlock,
 		reader:       backend.NewReader(l),
 		writer:       backend.NewWriter(l),
@@ -45,17 +46,17 @@ func NewLocalBlock(ctx context.Context, existingBlock common.BackendBlock, l *lo
 		}
 	}
 
-	return c, nil
+	return c
 }
 
-func (c *LocalBlock) FindTraceByID(ctx context.Context, id common.ID, opts common.SearchOptions) (*tempopb.Trace, error) {
+func (c *localBlock) FindTraceByID(ctx context.Context, id common.ID, opts common.SearchOptions) (*tempopb.Trace, error) {
 	return c.BackendBlock.FindTraceByID(ctx, id, opts)
 }
 
 // FlushedTime returns the time the block was flushed.  Will return 0
 //
 //	if the block was never flushed
-func (c *LocalBlock) FlushedTime() time.Time {
+func (c *localBlock) FlushedTime() time.Time {
 	unixTime := c.flushedTime.Load()
 	if unixTime == 0 {
 		return time.Time{} // return 0 time.  0 unix time is jan 1, 1970
@@ -63,7 +64,7 @@ func (c *LocalBlock) FlushedTime() time.Time {
 	return time.Unix(unixTime, 0)
 }
 
-func (c *LocalBlock) SetFlushed(ctx context.Context) error {
+func (c *localBlock) SetFlushed(ctx context.Context) error {
 	flushedTime := time.Now()
 	flushedBytes, err := flushedTime.MarshalText()
 	if err != nil {
@@ -79,7 +80,7 @@ func (c *LocalBlock) SetFlushed(ctx context.Context) error {
 	return nil
 }
 
-func (c *LocalBlock) Write(ctx context.Context, w backend.Writer) error {
+func (c *localBlock) Write(ctx context.Context, w backend.Writer) error {
 	err := encoding.CopyBlock(ctx, c.BlockMeta(), c.reader, w)
 	if err != nil {
 		return errors.Wrap(err, "error copying block from local to remote backend")