+++ tempodb/encoding/vparquet/block_findtracebyid.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"context"
 	"fmt"
+	"io"
 
 	"github.com/opentracing/opentracing-go"
 	"github.com/opentracing/opentracing-go/log"
@@ -69,7 +70,7 @@ func (b *backendBlock) FindTraceByID(ctx context.Context, traceID common.ID, opt
 		return nil, nil
 	}
 
-	pf, rr, err := b.openForSearch(derivedCtx, opts, parquet.SkipPageIndex(true))
+	pf, rr, err := b.openForSearch(derivedCtx, opts)
 	if err != nil {
 		return nil, fmt.Errorf("unexpected error opening parquet file: %w", err)
 	}
@@ -85,12 +86,15 @@ func (b *backendBlock) FindTraceByID(ctx context.Context, traceID common.ID, opt
 	}
 
 	numRowGroups := len(pf.RowGroups())
+	buf := make(parquet.Row, 1)
 
 	// Cache of row group bounds
 	rowGroupMins := make([]common.ID, numRowGroups+1)
 	rowGroupMins[0] = b.meta.MinID
 	rowGroupMins[numRowGroups] = b.meta.MaxID // This is actually inclusive and the logic is special for the last row group below
 
+	// Gets the minimum trace ID within the row group. Since the column is sorted
+	// ascending we just read the first value from the first page.
 	getRowGroupMin := func(rgIdx int) (common.ID, error) {
 		min := rowGroupMins[rgIdx]
 		if len(min) > 0 {
@@ -98,25 +102,24 @@ func (b *backendBlock) FindTraceByID(ctx context.Context, traceID common.ID, opt
 			return min, nil
 		}
 
-		// Read first value from the row group
-		rgs := pf.RowGroups()
-		iter := parquetquery.NewColumnIterator(derivedCtx, rgs[rgIdx:rgIdx+1], colIndex, "", 1, nil, "id")
-		defer iter.Close()
+		pages := pf.RowGroups()[rgIdx].ColumnChunks()[colIndex].Pages()
+		defer pages.Close()
 
-		res, err := iter.Next()
+		page, err := pages.ReadPage()
 		if err != nil {
 			return nil, err
 		}
-		if res == nil {
-			// This shouldn't happen as row groups are never empty, however have seen this in practice
-			// and it means there must be an error condition or some other situation that is not
-			// handled as expected.
-			return nil, fmt.Errorf("failed to read 1 value from row group: traceID: %s blockID:%v rowGroupIdx:%d", util.TraceIDToHexString(traceID), b.meta.BlockID, rgIdx)
+
+		c, err := page.Values().ReadValues(buf)
+		if err != nil && err != io.EOF {
+			return nil, err
+		}
+		if c < 1 {
+			return nil, fmt.Errorf("failed to read value from page: traceID: %s blockID:%v rowGroupIdx:%d", util.TraceIDToHexString(traceID), b.meta.BlockID, rgIdx)
 		}
 
-		min = res.ToMap()["id"][0].ByteArray()
+		min = buf[0].ByteArray()
 		rowGroupMins[rgIdx] = min
-
 		return min, nil
 	}